+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import './styles/variable.css'
-import { GlobalStyle } from './styles/GlobalStyle'
+import './styles/s-global/variable.css'
+import { GlobalStyle } from './styles/s-global/GlobalStyle'
 import Home from './pages/Home'
 import Join from './pages/Join'
 import Header from './components/Header'
