+++ client/src/components/Header.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { Wrapper, LogoContainer } from '../styles/header'
+import { Wrapper, LogoContainer } from '../styles/s-components/header'
 //* header에 있어야 할 기능
 // center - 로고 : 클릭하면 index로 이동
