+++ client/src/pages/Join.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { OuterWrapper } from '../styles/common'
+import { OuterWrapper } from '../styles/s-global/common'
 
 const Join = () => {
   return <OuterWrapper>Join</OuterWrapper>
