+++ client/src/styles/s-components/listcontent.js
