+++ client/src/styles/s-components/search.js
