+++ client/src/styles/s-global/variable.css
@@ -10,7 +10,7 @@
   --fz-big: 18px;
   --fz-base: 16px;
   --fz-sm: 14px;
-  --fz-xs: 10px;
+  --fz-xs: 12px;
 
   /* font weight */
   --fw-base: 300;
