+++ client/src/styles/s-pages/home.js
