+++ client/src/styles/s-pages/index.js
